Fixed SDK
Accidently uploaded a past version of sdk, fixed that while also adding patrick's code to test slides.
